Moved interfaces to same file as classes where only one implementation is expected.
