+++ backend/Application/Aggregates/Contract/ContractRepository.cs
@@ -9,6 +9,43 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.Aggregates.Contract;
 
+public interface IContractRepository : IAsyncDisposable
+{
+    /// <summary>
+    /// From block <see cref="heightFrom"/> to block <see cref="heightTo"/> all transaction result event
+    /// related to smart contract will be returned.
+    /// </summary>
+    Task<IList<TransactionResultEventDto>> FromBlockHeightRangeGetContractRelatedTransactionResultEventRelations(ulong heightFrom, ulong heightTo);
+    /// <summary>
+    /// From block <see cref="heightFrom"/> to block <see cref="heightTo"/> return all block heights
+    /// which has already been read and processed successfully.
+    /// </summary>
+    public Task<List<ulong>> FromBlockHeightRangeGetBlockHeightsReadOrdered(ulong heightFrom, ulong heightTo);
+    /// <summary>
+    /// Returns latest smart contract read height ordered descending by block height. 
+    /// </summary>
+    Task<ContractReadHeight?> GetReadOnlyLatestContractReadHeight();
+    /// <summary>
+    /// Get latest import state of block- and transactions ordered by
+    /// block slot time.
+    ///
+    /// Should return zero (default) is no entity is present.
+    /// </summary>
+    Task<long> GetReadOnlyLatestImportState(CancellationToken token = default);
+    /// <summary>
+    /// Adds entity to repository.
+    /// </summary>
+    Task AddAsync<T>(params T[] entities) where T : class;
+    /// <summary>
+    /// Adds entities to repository,
+    /// </summary>
+    Task AddRangeAsync<T>(IEnumerable<T> heights) where T : class;
+    /// <summary>
+    /// Save changes to storage.
+    /// </summary>
+    Task SaveChangesAsync(CancellationToken token);
+}
+
 internal sealed class ContractRepository : IContractRepository
 {
     private readonly GraphQlDbContext _context;
