--- backend/Application/Aggregates/Contract/IContractRepositoryFactory.cs
@@ -1,8 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Application.Aggregates.Contract;
-
-public interface IContractRepositoryFactory
-{
-    Task<IContractRepository> CreateAsync();
-}
