+++ backend/Application/Aggregates/Contract/Jobs/ContractJobFinder.cs
@@ -2,6 +2,11 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Application.Aggregates.Contract.Jobs;
 
+public interface IContractJobFinder
+{
+    IEnumerable<IContractJob> GetJobs();
+}
+
 public sealed class ContractJobFinder : IContractJobFinder
 {
     private readonly IServiceProvider _provider;
