--- backend/Application/Aggregates/Contract/Jobs/IContractJobFinder.cs
@@ -1,6 +1,0 @@
-namespace Application.Aggregates.Contract.Jobs;
-
-public interface IContractJobFinder
-{
-    IEnumerable<IContractJob> GetJobs();
-}
